+++ wookie/connectors/__init__.py
@@ -1,0 +1,5 @@
+# Module Description
+## This module is about connecting two sources (one called 'left', one called 'right')
+## And creating a side-by-side array of the different features
+
+
