+++ wookie/connectors/dataframes/__init__.py
@@ -1,0 +1,1 @@
+from wookie.connectors.dataframes.cartesianconnector import CartesianConnector
