+++ wookie/preutils.py
@@ -2,7 +2,6 @@ import unicodedata
 
 import numpy as np
 import pandas as pd
-from fuzzywuzzy.fuzz import ratio as simpleratio, partial_token_set_ratio as tokenratio
 
 suffixexact = 'exact'
 suffixtoken = 'token'
@@ -352,69 +351,3 @@ def datapasser(df):
         pd.DataFrame
     """
     return df
-
-
-def valid_inputs(left, right):
-    """
-    takes two inputs and return True if none of them is null, or False otherwise
-    Args:
-        left: first object (scalar)
-        right: second object (scalar)
-
-    Returns:
-        bool
-    """
-    if any(pd.isnull([left, right])):
-        return False
-    else:
-        return True
-
-
-def exact_score(left, right):
-    """
-    Checks if the two values are equali
-    Args:
-        left (object): object number 1
-        right (object): object number 2
-
-    Returns:
-        float
-    """
-    if valid_inputs(left, right) is False:
-        return navalue_score
-    else:
-        return float(left == right)
-
-
-def fuzzy_score(left, right):
-    """
-    return ratio score of fuzzywuzzy
-    Args:
-        left (str): string number 1
-        right (str): string number 2
-
-    Returns:
-        float
-    """
-    if valid_inputs(left, right) is False:
-        return navalue_score
-    else:
-        s = (simpleratio(left, right) / 100)
-    return s
-
-
-def token_score(left, right):
-    """
-    return the token_set_ratio score of fuzzywuzzy
-    Args:
-        left (str): string number 1
-        right (str): string number 2
-
-    Returns:
-        float
-    """
-    if valid_inputs(left, right) is False:
-        return navalue_score
-    else:
-        s = tokenratio(left, right) / 100
-    return s
